test: restore single integration spec and keep extracted helpers
